Docs: update create-a-store guide
* call out prefix and localKeys fields
* make the sample code block wider

The second part required removing some of the formatting. It honestly looked nicer before, but I think this is more functional.

I also looked for a way to have the sample code in it's own standalone .ts and .js files, but couldn't figure out a way to do that and automatically import them into the document, and I think having example code in this document is more valuable. Reducing the level of indentation at least helped.