Default to /56 subnet block for IPv6
This enables blocking of entire subnets for IPv6 with a somewhat aggressive default block of /56.

It also adds a helper function that can be used as a fallback in custom keyGenerators, and a couple of new validation checks.

Todo:
* [ ] Readme example
* [ ] Readme config option
* [ ] Docs for new config option
* [ ] Docs for new error codes
* [ ] Redirects for new error codes (different repo)

I also added the key to the `key` to the `RateLimitInfo` object since it's no longer quite as obvious.