fix: export ipKeyGenerator in cjs build
We add a footer to the cjs that redefines the exports in order to keep the original import style working, but it had been hard-coded to specific fields rather than just re-exporting everything. Now it re-exports everything, so we shouldn't need to worry about it again if we add more exports in the future.

Fixes #529