fix: replace vulnerable ip package
This drops the `ip` library that v8.0.0 required due to an open vulnerability in the library.
express-rate-limit did not use the vulnerable code, but it still caused warnings and reports for end users.

The replacement is a much smaller and more focused bit of code generated by Google Gemini that handles the ip bit fiddling we need to apply the subnet mask, but not much else.

The code looks correct to me, and passes all of our tests (including the ones where supertest sets the source IP to `::ffff:127.0.0.1`, and the tests that gemini wrote to go with it.

That said, please review with extra scrutiny since this code wasn't written by a human.