test: check for renamed Request in types
This is an incredibly dumb test, but I couldn't come up with anything smarter that actually worked - it all either failed in unexpected ways, or passed even when the types were broken.

I tried a couple of differnt ways to import and use the types in the test, and I also tried a couple of different type linting sort of tools.

Relates to:
* https://github.com/express-rate-limit/express-rate-limit/issues/530
* https://github.com/express-rate-limit/express-rate-limit/pull/534
* https://github.com/express-rate-limit/express-rate-limit/pull/537#issuecomment-3099816379
* https://github.com/timocov/dts-bundle-generator/issues/353