Add validation check for Forwarded header
This alerts the user if a Forwarded header (standardized version of X-Forwarded-For & friends) is set when using the default keyGenerator, similar to our check for X-Forwarded-For when `trust proxy` is not enabled.

I also added an example to the docs with a custom keyGenerator that uses the [forwarded-parse](https://www.npmjs.com/package/forwarded-parse) library to parse and use the rightmost forwarded header (e.g. the last one, from the nearest proxy server).